+++ src/miniprogram/group/group.ts
@@ -33,33 +33,19 @@ import {
   setGroupVerificationParams,
   transferGroupOwnerParams,
 } from "./params";
-import { createWorker, stopWorker } from "open-im-sdk/util";
+import axios from "axios";
 
 export default class OpenIMSDK extends Emitter {
-  private ws: WebSocket | undefined;
   private uid: string | undefined;
   private token: string | undefined;
-  private platform: string = "miniProgram";
-  private wsUrl: string = "";
-  private lock: boolean = false;
-  private logoutFlag: boolean = false;
-  private ws2promise: Record<string, Ws2Promise> = {};
-  private onceFlag: boolean = true;
-  private timer: NodeJS.Timer | undefined = undefined;
-  private lastTime: number = 0;
-  private heartbeatCount: number = 0;
-  private heartbeatStartTime: number = 0;
-  private platformID: number = 0;
-  private isBatch: boolean = false;
-  private worker: Worker | null = null;
-
+  private baseUrl: string = "http://localhost:10002";
   constructor() {
     super();
-    this.getPlatform();
   }
 
   // accept group application
   acceptGroupApplication = (
+    url: string = this.baseUrl,
     data: acceptGroupApplicationParams,
     operationID: string
   ) => {
@@ -70,12 +56,13 @@ export default class OpenIMSDK extends Emitter {
         operationID: operationID,
         userID: this.uid,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // change GroupMember to Mute
   changeGroupMemberMute = (
+    url: string = this.baseUrl,
     data: ChangeGroupMemberMuteParams,
     operationID: string
   ) => {
@@ -86,12 +73,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.CHANGEGROUPMEMBERMUTE,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // changeGroup to Mute
-  changeGroupToMute = (data: changeGroupToMuteParams, operationID: string) => {
+  changeGroupToMute = (
+    url: string = this.baseUrl,
+    data: changeGroupToMuteParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -99,12 +90,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.CHANGEGROUPTOMUTE,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // create a Group
-  createGroup = (data: changeGroupToMuteParams, operationID: string) => {
+  createGroup = (
+    url: string = this.baseUrl,
+    data: changeGroupToMuteParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -112,12 +107,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.CREATEGROUP,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // dismiss a group
-  dismissGroup = (data: dismissGroupParams, operationID: string) => {
+  dismissGroup = (
+    url: string = this.baseUrl,
+    data: dismissGroupParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -125,12 +124,15 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.DISMISSGROUP,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // get Group ApplicationList As Applicant
-  getGroupApplicationListAsApplicant = (operationID: string) => {
+  getGroupApplicationListAsApplicant = (
+    url: string = this.baseUrl,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         operationID: operationID,
@@ -138,12 +140,15 @@ export default class OpenIMSDK extends Emitter {
         reqFuncName: RequestFunc.GETGROUPAPPLICATIONLISTASAPPLICANT,
         data: "",
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // getGroupApplicationList As Recipient
-  getGroupApplicationListAsRecipient = (operationID: string) => {
+  getGroupApplicationListAsRecipient = (
+    url: string = this.baseUrl,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         operationID: operationID,
@@ -151,12 +156,13 @@ export default class OpenIMSDK extends Emitter {
         reqFuncName: RequestFunc.GETGROUPAPPLICATIONLISTASRECIPIENT,
         data: "",
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // getGroupMemberList
   getGroupMemberList = (
+    url: string = this.baseUrl,
     data: getGroupMemberListParams,
     operationID: string
   ) => {
@@ -167,12 +173,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.GETGROUPMEMBERLIST,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // getGroupMemberListByJoinTimeFilter
   getGroupMemberListByJoinTimeFilter = (
+    url: string = this.baseUrl,
     data: getGroupMemberListByJoinTimeFilterParams,
     operationID: string
   ) => {
@@ -183,12 +190,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.GETGROUPMEMBERLISTBYJOINTIMEFILTER,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // getGroupMemberOwnerAndAdmin
   getGroupMemberOwnerAndAdmin = (
+    url: string = this.baseUrl,
     data: getGroupMemberOwnerAndAdminParams,
     operationID: string
   ) => {
@@ -199,12 +207,12 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.GETGROUPMEMBEROWNERANDADMIN,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // get JoinedGroupList
-  getJoinedGroupList = (operationID: string) => {
+  getJoinedGroupList = (url: string = this.baseUrl, operationID: string) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         operationID: operationID,
@@ -212,12 +220,13 @@ export default class OpenIMSDK extends Emitter {
         reqFuncName: RequestFunc.GETJOINEDGROUPLIST,
         data: "",
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // get SpecifiedGroupMembersInfo
   getSpecifiedGroupMembersInfo = (
+    url: string = this.baseUrl,
     data: getSpecifiedGroupMembersInfoParams,
     operationID: string
   ) => {
@@ -228,12 +237,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.GETSPECIFIEDGROUPMEMBERSINFO,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // get SpecifiedGroupsInfo
   getSpecifiedGroupsInfo = (
+    url: string = this.baseUrl,
     data: getSpecifiedGroupsInfoParams,
     operationID: string
   ) => {
@@ -244,12 +254,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.GETSPECIFIEDGROUPSINFO,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // inviteUser To Group
-  inviteUserToGroup = (data: inviteUserToGroupParams, operationID: string) => {
+  inviteUserToGroup = (
+    url: string = this.baseUrl,
+    data: inviteUserToGroupParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -257,12 +271,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.INVITEUSERTOGROUP,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // joinGroup
-  joinGroup = (data: joinGroupParams, operationID: string) => {
+  joinGroup = (
+    url: string = this.baseUrl,
+    data: joinGroupParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -270,12 +288,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.JOINGROUP,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // kick GroupMember
-  kickGroupMember = (data: kickGroupMemberParams, operationID: string) => {
+  kickGroupMember = (
+    url: string = this.baseUrl,
+    data: kickGroupMemberParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -283,12 +305,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.KICKGROUPMEMBER,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // quitGroup
-  quitGroup = (data: kickGroupMemberParams, operationID: string) => {
+  quitGroup = (
+    url: string = this.baseUrl,
+    data: kickGroupMemberParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -296,12 +322,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.QUITGROUP,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // refuse a  GroupApplication
   refuseGroupApplication = (
+    url: string = this.baseUrl,
     data: refuseGroupApplicationParams,
     operationID: string
   ) => {
@@ -312,12 +339,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.REFUSEGROUPAPPLICATION,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // searchGroupMembers
   searchGroupMembers = (
+    url: string = this.baseUrl,
     data: searchGroupMembersParams,
     operationID: string
   ) => {
@@ -328,12 +356,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SEARCHGROUPMEMBERS,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // searchGroups
-  searchGroups = (data: searchGroupsParams, operationID: string) => {
+  searchGroups = (
+    url: string = this.baseUrl,
+    data: searchGroupsParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -341,12 +373,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SEARCHGROUPS,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   //set GroupApplyMemberFriend
   setGroupApplyMemberFriend = (
+    url: string = this.baseUrl,
     data: setGroupApplyMemberFriendParams,
     operationID: string
   ) => {
@@ -357,12 +390,16 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPAPPLYMEMBERFRIEND,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // setGroupInfo
-  setGroupInfo = (data: setGroupInfoParams, operationID: string) => {
+  setGroupInfo = (
+    url: string = this.baseUrl,
+    data: setGroupInfoParams,
+    operationID: string
+  ) => {
     return new Promise<WsResponse>((resolve, reject) => {
       const args = {
         data: data,
@@ -370,12 +407,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPINFO,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // setGroupLookMemberInfo
   setGroupLookMemberInfo = (
+    url: string = this.baseUrl,
     data: setGroupLookMemberInfoParams,
     operationID: string
   ) => {
@@ -386,12 +424,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPLOOKMEMBERINFO,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // setGroupMemberInfo
   setGroupMemberInfo = (
+    url: string = this.baseUrl,
     data: setGroupMemberInfoParams,
     operationID: string
   ) => {
@@ -402,12 +441,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPMEMBERINFO,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // setGroupMemberNickname
   setGroupMemberNickname = (
+    url: string = this.baseUrl,
     data: setGroupMemberNicknameParams,
     operationID: string
   ) => {
@@ -418,12 +458,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPMEMBERNICKNAME,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // setGroupMemberRoleLevel
   setGroupMemberRoleLevel = (
+    url: string = this.baseUrl,
     data: setGroupMemberRoleLevelParams,
     operationID: string
   ) => {
@@ -434,12 +475,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPMEMBERROLELEVEL,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // setGroupVerification
   setGroupVerification = (
+    url: string = this.baseUrl,
     data: setGroupVerificationParams,
     operationID: string
   ) => {
@@ -450,12 +492,13 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.SETGROUPVERIFICATION,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // transferGroupOwner
   transferGroupOwner = (
+    url: string = this.baseUrl,
     data: transferGroupOwnerParams,
     operationID: string
   ) => {
@@ -466,329 +509,31 @@ export default class OpenIMSDK extends Emitter {
         userID: this.uid,
         reqFuncName: RequestFunc.TRANSFERGROUPOWNER,
       };
-      this.wsSend(args, resolve, reject);
+      this.HttpSend(args, url);
     });
   };
 
   // connect tools
-  private wsSend = (
-    params: WsParams,
-    resolve: (value: WsResponse | PromiseLike<WsResponse>) => void,
-    reject: (reason?: any) => void
-  ) => {
-    if (window?.navigator && !window.navigator.onLine) {
-      let errData: WsResponse = {
-        event: params.reqFuncName,
-        errCode: 113,
-        errMsg: "net work error",
-        data: "",
-        operationID: params.operationID || "",
-      };
-      reject(errData);
-      return;
-    }
-    if (this.ws?.readyState !== this.ws?.OPEN) {
-      let errData: WsResponse = {
-        event: params.reqFuncName,
-        errCode: 112,
-        errMsg: "ws conecting...",
-        data: "",
-        operationID: params.operationID || "",
-      };
-      reject(errData);
-      return;
-    }
-    // if (this.ws === undefined) {
-    //   let errData: WsResponse = {
-    //     event: params.reqFuncName,
-    //     errCode: 112,
-    //     errMsg: "ws conect failed...",
-    //     data: "",
-    //     operationID: params.operationID || "",
-    //   };
-    //   reject(errData);
-    //   return;
-    // }
-    if (typeof params.data === "object") {
-      params.data = JSON.stringify(params.data);
-    }
-
-    const ws2p = {
-      oid: params.operationID || uuid(this.uid as string),
-      mname: params.reqFuncName,
-      mrsve: resolve,
-      mrjet: reject,
-      flag: false,
-    };
-
-    this.ws2promise[ws2p.oid] = ws2p;
-
-    const handleMessage = (ev: MessageEvent<string>) => {
-      this.lastTime = new Date().getTime();
-      const data = JSON.parse(ev.data);
-      if ((CbEvents as Record<string, string>)[data.event.toUpperCase()]) {
-        this.emit(data.event, data);
-        return;
-      }
-
-      if (params.reqFuncName === RequestFunc.LOGOUT) {
-        this.logoutFlag = true;
-        this.ws!.close();
-        this.ws = undefined;
-      }
-
-      const callbackJob = this.ws2promise[data.operationID];
-      if (!callbackJob) {
-        if (
-          data.event === RequestFunc.SENDMESSAGE ||
-          data.event === RequestFunc.SENDMESSAGENOTOSS
-        ) {
-          this.emit(CbEvents.ONRECVNEWMESSAGEFROMOTHERWEB, data);
-        }
-        return;
-      }
-      if (data.errCode === 0) {
-        callbackJob.mrsve(data);
-      } else {
-        callbackJob.mrjet(data);
-      }
-      delete this.ws2promise[data.operationID];
-    };
-
-    try {
-      if (this.platform == "web") {
-        this.ws!.send(JSON.stringify(params));
-        this.ws!.onmessage = handleMessage;
-      } else {
-        this.ws!.send({
-          //@ts-ignore
-          data: JSON.stringify(params),
-          success: (res: any) => {
-            //@ts-ignore
-            if (
-              this.platform === "uni" &&
-              //@ts-ignore
-              this.ws!._callbacks !== undefined &&
-              //@ts-ignore
-              this.ws!._callbacks.message !== undefined
-            ) {
-              //@ts-ignore
-              this.ws!._callbacks.message = [];
-            }
-          },
-        });
-        if (this.onceFlag) {
-          //@ts-ignore
-          this.ws!.onMessage(handleMessage);
-          this.onceFlag = false;
-        }
-      }
-    } catch (error) {
-      let errData: WsResponse = {
-        event: params.reqFuncName,
-        errCode: 112,
-        errMsg: "no ws conect...",
-        data: "",
-        operationID: params.operationID || "",
-      };
-      reject(errData);
-      return;
-    }
-    if (params.reqFuncName === RequestFunc.LOGOUT) {
-      this.onceFlag = true;
-    }
-  };
-
-  // get login credentials
-  private iLogin(data: LoginParams, operationID?: string) {
-    return new Promise<WsResponse>((resolve, reject) => {
-      const _uuid = operationID || uuid(this.uid as string);
-      const args = {
-        reqFuncName: RequestFunc.LOGIN,
-        operationID: _uuid,
-        userID: this.uid,
-        data,
-        batchMsg: this.isBatch ? 1 : 0,
-      };
-      this.wsSend(args, resolve, reject);
-    });
-  }
-
-  // user logo out
-  logout(operationID?: string) {
-    return new Promise<WsResponse>((resolve, reject) => {
-      const _uuid = operationID || uuid(this.uid as string);
-      const args = {
-        reqFuncName: RequestFunc.LOGOUT,
-        operationID: _uuid,
-        userID: this.uid,
-        data: "",
-      };
-      this.wsSend(args, resolve, reject);
-    });
-  }
-  // get login status
-  getLoginStatus = (operationID?: string) => {
-    return new Promise<WsResponse>((resolve, reject) => {
-      const _uuid = operationID || uuid(this.uid as string);
-      const args = {
-        reqFuncName: RequestFunc.GETLOGINSTATUS,
-        operationID: _uuid,
-        userID: this.uid,
-        data: "",
-      };
-      this.wsSend(args, resolve, reject);
-    });
-  };
-
-  private getPlatform() {
-    const wflag = typeof WebSocket;
-    //@ts-ignore
-    const uflag = typeof uni;
-    //@ts-ignore
-    const xflag = typeof wx;
-    //@ts-ignore
-    const mflag = typeof miniprogram;
-
-    if (wflag !== "undefined") {
-      this.platform = "web";
-      return;
-    }
-
-    if (uflag === "object" && xflag !== "object" && mflag !== "object") {
-      this.platform = "uni";
-    } else if (uflag !== "object" && xflag === "object" && mflag !== "object") {
-      this.platform = "wx";
-    } else if (uflag !== "object" && xflag === "object" && mflag !== "object") {
-      this.platform = "miniProgram";
-    } else {
-      this.platform = "unknow";
-    }
-  }
-
-  private createWs(
-    _onOpen?: Function,
-    _onClose?: Function,
-    _onError?: Function
-  ) {
-    console.log("start createWs...");
-    return new Promise<void>((resolve, reject) => {
-      this.ws?.close();
-      this.ws = undefined;
-
-      let onOpen: any = () => {
-        const loginData = {
-          userID: this.uid!,
-          token: this.token!,
+  private HttpSend = (params: WsParams, url: string) => {
+    return new Promise((resolve, reject) => {
+      if (window?.navigator && !window.navigator.onLine) {
+        let errData = {
+          event: params.reqFuncName,
+          errCode: 113,
+          data: "",
+          operationID: params.operationID || "",
         };
-        this.iLogin(loginData).then((res) => {
-          this.logoutFlag = false;
-          console.log("iLogin suc...");
-          this.heartbeat();
-          resolve();
+        reject(errData);
+        return;
+      }
+      axios
+        .post(url, params)
+        .then((response) => {
+          resolve(response.data);
+        })
+        .catch((error) => {
+          reject(error);
         });
-      };
-
-      if (_onOpen) {
-        onOpen = _onOpen;
-      }
-
-      let onClose: any = () => {
-        console.log("ws close agin:::");
-        if (!this.logoutFlag) {
-          Object.values(this.ws2promise).forEach((promise) =>
-            promise.mrjet({
-              event: promise.mname,
-              errCode: 111,
-              errMsg: "ws connect close...",
-              data: "",
-              operationID: promise.oid,
-            })
-          );
-          // this.reconnect();
-        }
-      };
-
-      if (_onClose) {
-        onClose = _onClose;
-      }
-
-      let onError: any = () => {};
-      if (_onError) {
-        onError = _onError;
-      }
-
-      if (this.platform === "miniProgram") {
-        this.ws = new WebSocket(this.wsUrl);
-        this.ws.onclose = onClose;
-        this.ws.onopen = onOpen;
-        this.ws.onerror = onError;
-        return;
-      }
-
-      // @ts-ignore
-      const platformNamespace = this.platform === "uni" ? uni : wx;
-      this.ws = platformNamespace.connectSocket({
-        url: this.wsUrl,
-        complete: () => {},
-      });
-      //@ts-ignore
-      this.ws.onClose(onClose);
-      //@ts-ignore
-      this.ws.onOpen(onOpen);
-      //@ts-ignore
-      this.ws.onError(onError);
     });
-  }
-
-  private reconnect() {
-    if (!this.onceFlag) this.onceFlag = true;
-    if (this.lock) return;
-    this.lock = true;
-    this.clearTimer();
-    this.timer = setTimeout(() => {
-      this.createWs();
-      this.lock = false;
-    }, 500);
-  }
-  // cleartimer
-  private clearTimer() {
-    if (this.timer) {
-      clearTimeout(this.timer);
-    }
-  }
-  // test heartbeat
-  private heartbeat() {
-    console.log("start heartbeat...");
-    this.clearTimer();
-    const callback = () => {
-      if (this.logoutFlag) {
-        if (this.worker) {
-          stopWorker(this.worker);
-        }
-        return;
-      }
-
-      if (
-        this.ws?.readyState !== this.ws?.CONNECTING &&
-        this.ws?.readyState !== this.ws?.OPEN
-      ) {
-        this.reconnect();
-        return;
-      }
-
-      const now = new Date().getTime();
-      if (now - this.lastTime < 9000) {
-        return;
-      }
-
-      this.getLoginStatus().catch((err) => this.reconnect());
-    };
-    if (this.worker) {
-      stopWorker(this.worker);
-    }
-    try {
-      this.worker = createWorker(callback, 10000);
-    } catch (error) {}
-  }
+  };
 }
